Issue #20 - update Makefile. Made database initialization a part of the normal programs main method. Updated documentation. Split Minitwit.go into a few more modules
