+++ Makefile
@@ -1,8 +1,17 @@
-init:
-	python3 -c"from minitwit import init_db; init_db()"
-
 build:
-	gcc flag_tool.c -l sqlite3 -L/opt/local/lib/ -o flag_tool -g
+	go build src/minitwit.go
+	gcc src/flag_tool.c -l sqlite3 -L/opt/local/lib/ -o src/flag_tool -g
+
+start:
+	go run src/minitwit.go
 
 clean:
-	rm flag_tool
+	rm src/flag_tool
+	rm src/minitwit
+
+
+inspectdb:
+	./src/flag_tool -i | less
+
+flag:
+	./src/flag_tool "$@"
