+++ src/static/templates/login.html
@@ -1,0 +1,42 @@
+<!doctype html>
+<title>Welcome| MiniTwit</title>
+<head>
+  <link rel=stylesheet type=text/css href="/css/style.css">
+</head>
+
+<body>
+<div class=page>
+  <h1>MiniTwit</h1>
+  <div class=navigation>
+{{ if false }}
+    <a href="">my timeline</a> |
+    <a href="/public" >public timeline</a> |
+    <a href="/logout">sign out USERNAME</a>
+{{ else if true }}
+    <a href="/public">public timeline</a> |
+    <a href="/register">sign up</a> |
+    <a href="/login">sign in</a href="/login">
+{{end}}
+  </div>
+
+
+  <div class=body>
+  <h2>Sign In</h2>
+  {{ if .HasError }}<div class=error><strong>Error:</strong> {{ .ErrorMsg }}</div>{{ end }}
+  <form action="" method=post>
+    <dl>
+      <dt>Username:
+      <dd><input type=text name=username size=30 value="">
+      <dt>Password:
+      <dd><input type=password name=password size=30>
+    </dl>
+    <div class=actions><input type=submit value="Sign In"></div>
+  </form>
+</div>
+
+<div class=footer>
+  MiniTwit &mdash; A Golang Application
+</div>
+</div>
+</body>
+
