+++ src/static/templates/timeline.html
@@ -1,0 +1,91 @@
+<!doctype html>
+<title>Welcome| MiniTwit</title>
+<head>
+  <link rel=stylesheet type=text/css href="/css/style.css">
+
+</head>
+<body>
+<div class=page>
+  <h1>MiniTwit</h1>
+  <div class=navigation>
+{{ if false }}
+    <a href="">my timeline</a> |
+    <a href="/public" >public timeline</a> |
+    <a href="/logout">sign out USERNAME</a>
+{{ else if true }}
+    <a href="/public">public timeline</a> |
+    <a href="/register">sign up</a> |
+    <a href="/login">sign in</a href="/login">
+{{end}}
+  </div>
+  
+  <div class=body>
+
+
+
+
+<h2>
+  {{ if eq "public_timeline" .RequestEndpoint }}
+    Public Timeline
+  {{ else if eq "user_timeline" .RequestEndpoint}}
+    {{ .username }}'s Timeline
+  {{ else }}
+    My Timeline
+  {{ end }}
+ </h2>
+
+{{/*
+  <h2>{{ self.title() }}</h2>
+  {% if g.user %}
+    {% if request.endpoint == 'user_timeline' %}
+      <div class=followstatus>
+      {% if g.user.user_id == profile_user.user_id %}
+        This is you!
+      {% elif followed %}
+        You are currently following this user.
+        <a class=unfollow href="{{ url_for('unfollow_user', username=profile_user.username)
+          }}">Unfollow user</a>.
+      {% else %}
+        You are not yet following this user.
+        <a class=follow href="{{ url_for('follow_user', username=profile_user.username)
+          }}">Follow user</a>.
+      {% endif %}
+      </div>
+    {% elif request.endpoint == 'timeline' %}
+      <div class=twitbox>
+        <h3>What's on your mind {{ g.user.username }}?</h3>
+        <form action="{{ url_for('add_message') }}" method=post>
+          <p><input type=text name=text size=60><!--
+          --><input type=submit value="Share">
+        </form>
+      </div>
+    {% endif %}
+  {% endif %}
+*/}}
+  
+
+  <ul class=messages>
+    {{ if gt (len .Messages) 0}}
+      {{range .Messages}}
+        <li>
+          {{/* <img src={{.gravatar_url}}> */}}
+          <strong>
+            {{/*<a href="{{ url_for('user_timeline', username=message.username)}}">{{ message.username }}</a>*/}}
+            <a href="">{{ .Username }}</a>
+          </strong>
+          {{ .Text }}
+          <small>&mdash; {{ .Pub_date}}</small>
+        </li>
+      {{end}}
+    {{else}}
+    <em>There's no message so far.</em>
+    {{end}}
+  </ul>
+
+
+</div>
+<div class=footer>
+  MiniTwit &mdash; A Golang Application
+</div>
+</div>
+</body>
