+++ src/utils/utils.go
@@ -1,0 +1,13 @@
+package utils
+
+import "os"
+
+// Exists reports whether the named file or directory exists.
+func FileExists(name string) bool {
+	if _, err := os.Stat(name); err != nil {
+		if os.IsNotExist(err) {
+			return false
+		}
+	}
+	return true
+}
